examples: add bpf_core_read.h header and example using CO-RE helpers
This commit modifies `update.sh` to extract also the bpf_core_read.h
header file. Then, it adds a new `fentry` to showcase how `BPF_CORE_READ`
can be used to read fields from kernel structs without the entire struct
definition. That is, `BPF_CORE_READ` takes care of finding the offset of
the field the caller is looking for within the struct.